fix: renamed files to match architechture of the app
