+++ App.js
@@ -1,4 +1,4 @@
-import { Map, UserLogin, UserSignUp, Welcome } from './components';
+import { Main, UserLogin, UserSignUp, Welcome } from './components';
 import { createSwitchNavigator } from 'react-navigation';
 
 const App = createSwitchNavigator(
@@ -6,7 +6,7 @@ const App = createSwitchNavigator(
     Welcome,
     UserLogin,
     UserSignUp,
-    Map
+    Main
   }, {
     initialRouteName: 'Welcome'
   }
