+++ components/UserLogIn.js
@@ -21,7 +21,7 @@ class UserLogin extends Component {
             await firebase.initializeApp(firebaseConfig);
         }
         await firebase.auth().signInWithEmailAndPassword(this.state.email, this.state.password);
-        this.props.navigation.navigate('Map');
+        this.props.navigation.navigate('Main');
     } catch (err) {
         Alert.alert(err.toString())
     }
