+++ components/index.js
@@ -1,5 +1,5 @@
-export { Map } from './Map';
-export { default as UserLogin } from './UserLogIn';
 export { default as Main } from './Main';
+export { default as UserLogin } from './UserLogIn';
+export { default as Cards } from './Cards';
 export { default as UserSignUp } from './UserSignUp';
 export { default as Welcome } from './Welcome';
